[#71] 장바구니 예외처리

+++ src/main/java/flab/delideli/dto/CartDTO.java
@@ -10,5 +10,5 @@ import java.util.List;
 public class CartDTO {
 
 	List<CartlistDTO> cartlistDTOS;
-	int totalPrice;
+	long totalPrice;
 }
